fix(deps): 間接依存関係を更新し clippy 警告を修正
- cargo update で 99 件の間接依存関係を最新版に更新
- BigDecimal 比較での cmp_owned 警告を修正
  - 整数との比較で BigDecimal::from() を使わず直接比較するよう変更
  - BigDecimal::sign() と BigDecimal::abs() を使用

+++ backend/src/trade/predict/tests.rs
@@ -134,10 +134,7 @@ async fn test_get_top_tokens_with_specific_volatility() -> Result<()> {
             !token.token.to_string().is_empty(),
             "Token name should not be empty"
         );
-        assert!(
-            token.volatility >= BigDecimal::from(0),
-            "Volatility should be non-negative"
-        );
+        assert!(token.volatility >= 0, "Volatility should be non-negative");
     }
 
     for i in 1..tokens.len() {
